+++ search-service/src/main/java/com/uns/ftn/searchservice/dto/VehicleDTO.java
@@ -15,7 +15,7 @@ public class VehicleDTO {
     private Long id;
 
     @JsonProperty("kilometersTraveled")
-    private double kilometersTraveled;
+    private int kilometersTraveled;
 
     @JsonProperty("childSeatNumber")
     private int childSeatNumber;
